backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.0
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
passlib[bcrypt]===1.7.4
python-jose[cryptography]===3.3.0
python-multipart==0.0.9
pydantic[email]==1.10.13